Remove separate search types
