+++ src/pages/public_registration/RegistrationSummary.tsx
@@ -4,8 +4,8 @@ import { Skeleton } from '@mui/material';
 import { Link as RouterLink } from 'react-router-dom';
 import { useSearchRegistrations } from '../../utils/hooks/useSearchRegistrations';
 import { getRegistrationLandingPagePath } from '../../utils/urlPaths';
-import { SearchFieldName } from '../../types/search.types';
 import { getRegistrationIdentifier } from '../../utils/registration-helpers';
+import { RegistrationFieldName } from '../../types/publicationFieldNames';
 
 interface RegistrationSummaryProps {
   id: string;
@@ -14,7 +14,7 @@ interface RegistrationSummaryProps {
 export const RegistrationSummary = ({ id }: RegistrationSummaryProps) => {
   const identifier = getRegistrationIdentifier(id);
   const [searchContainer, isLoadingSearchContainer] = useSearchRegistrations({
-    properties: [{ fieldName: SearchFieldName.Id, value: identifier }],
+    properties: [{ fieldName: RegistrationFieldName.Id, value: identifier }],
   });
 
   const container = searchContainer && searchContainer.hits.length === 1 ? searchContainer.hits[0] : null;
