+++ src/pages/search/SearchResults.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 import { Divider, Typography } from '@mui/material';
-import { SearchResult } from '../../types/search.types';
 import { RegistrationList } from '../../components/RegistrationList';
+import { SearchResult } from '../../types/registration.types';
 
 const StyledSearchResults = styled.div`
   padding-bottom: 1rem;
